8.4 정리 중

+++ src/main/java/travelator/HasPrice.kt
@@ -1,5 +1,5 @@
 package travelator
 
 interface HasPrice {
-    val price: Double
+    val price: Int
 }
